backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
greenlet>=3.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
langchain-google-vertexai>=1.0.1
groq==0.4.0
chromadb==0.4.18
sentence-transformers>=3.0.0
scikit-learn==1.3.2
openai==1.3.7

# Data Processing
pandas==2.1.3
numpy==1.26.2

# File Processing
pypdf==3.17.0
pdfplumber==0.11.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
Pillow==10.1.0
python-pptx==0.6.23
python-docx==1.1.0
docx2txt==0.9

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0
bcrypt==4.1.2
